Rename "buffer" to "sampling"
