+++ piker/_daemon.py
@@ -137,7 +137,7 @@ _data_mods = [
     'piker.brokers.core',
     'piker.brokers.data',
     'piker.data',
-    'piker.data._buffer'
+    'piker.data._sampling'
 ]
