+++ piker/data/_sampling.py
